Update C++ standards and improve plotting functions
- Changed C++ language standard in `AudioAnalyzer.jucer` to "latest" and updated `AudioAnalyzer_App.vcxproj` to "stdcpplatest".
- Modified several binary files, including icons.
- In `SpectrogramComponent.cpp`, updated plotting logic to use `updatePlot` instead of `updatePlotRealTime`.
- In `cmp_utils.h`, added explicit float casting for logarithmic scaling return values to enhance type safety.

+++ Source/cmp_utils.h
@@ -124,7 +124,7 @@ constexpr float getXDataFromXPixelCoordinate(
       return coordinateToXLinear();
       break;
     case Scaling::logarithmic:
-      return coordinateToXLog();
+      return (float)(coordinateToXLog());
       break;
     default:
       return coordinateToXLinear();
@@ -152,7 +152,7 @@ constexpr float getYDataFromYPixelCoordinate(
       return coordinateToYLinear();
       break;
     case Scaling::logarithmic:
-      return coordinateToYLog();
+      return (float)(coordinateToYLog());
       break;
     default:
       return coordinateToYLinear();
